misc/builder: support for uploading tarballs to googlecode
R=rsc
CC=golang-dev
https://golang.org/cl/1784042

+++ src/clean.bash
@@ -20,7 +20,9 @@ fi
 
 GOBIN="${GOBIN:-$HOME/bin}"
 
-rm -rf "$GOROOT"/pkg/${GOOS}_$GOARCH
+if [ "$1" != "--nopkg" ]; then
+	rm -rf "$GOROOT"/pkg/${GOOS}_$GOARCH
+fi
 rm -f "$GOROOT"/lib/*.a
 for i in lib9 libbio libcgo libmach cmd pkg \
 	../misc/cgo/gmp ../misc/cgo/stdio \
